make rendering, raising of errors optional
Rendering here can cause frustrating DoubleRenderError exceptions, plus some applications may want to handle the edge cases differently.  This change preserves the old behavior by default and adds two keyword arguments: `render_errors` controls whether or not text is rendered when the requested range is invalid and `raise_errors` controls whether or not an exception should be raised instead, in which case a generic `RangeError` is raised.

Although it is technically not an error, removing all instances of "render" from this gem is desirable, so the `render_errors` option also applies to the "no content" case as well.